Add team points ranking
