+++ app/docker-compose.yml
@@ -25,6 +25,5 @@ services:
           - ./sql/9_insert_trains.sql:/docker-entrypoint-initdb.d/9_insert_trains.sql
           - ./sql/a_10_insert_championship.sql:/docker-entrypoint-initdb.d/a_10_insert_championship.sql
           - ./sql/a_15_insert_finance.sql:/docker-entrypoint-initdb.d/a_15_insert_finance.sql
-          - ./sql/a_16_insert_transaction.sql:/docker-entrypoint-initdb.d/a_16_insert_transaction.sql
           - ./sql/a_17_procedures.sql:/docker-entrypoint-initdb.d/a_17_procedures.sql
           - ./sql/init.sql:/docker-entrypoint-initdb.d/init.sql
